jobyte-data-analysis/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==5.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.1.3
requests==2.31.0
python-dotenv==1.2.1
gunicorn==23.0.0